move std-mesh-textured
